refactor: update RenderUtil to use Float for position and size parameters

+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/client/render/DiviningRodRenderer.kt
@@ -100,12 +100,12 @@ object DiviningRodRenderer {
 			RenderUtil.renderCube(
 				poseStack,
 				buffer,
-				indicator.target.x,
-				indicator.target.y,
-				indicator.target.z,
-				1,
-				1,
-				1,
+				indicator.target.x.toFloat(),
+				indicator.target.y.toFloat(),
+				indicator.target.z.toFloat(),
+				1f,
+				1f,
+				1f,
 				indicator.color
 			)
 		}
